rename to components, get rid of unecesasry table column files
