+++ app/src/admin/AdminHome/AdminTransactionsTable.tsx
@@ -1,16 +1,93 @@
-import { GridOverlay } from '@mui/x-data-grid';
+import { Chip } from '@mui/material';
+import { GridColDef, GridOverlay } from '@mui/x-data-grid';
 import { BallTriangle } from 'react-loader-spinner';
 
 import StyledDataGrid from '../components/StyledDataGrid';
 import CustomToolbar from '../DataGrid/CustomToolbar';
 import NoResultsOverlay from '../DataGrid/NoResultsOverlay';
-import { Transaction } from '../types';
-import { generateTransactionColumns } from './transactionColumns';
+import { Transaction, TransactionType } from '../types';
+import { capitalizeFirstLetter, tagColor } from '../utils';
 
 interface Props {
     filteredRows: Transaction[];
     loadingTransactions: boolean;
     dishTypes: string[];
+}
+
+export function generateTransactionColumns(dishTypes: string[]): GridColDef[] {
+    return [
+        { field: 'qid', headerName: 'Dish Id', minWidth: 150, maxWidth: 150, flex: 1 },
+        {
+            field: 'dishType',
+            headerName: 'Dish Type',
+            minWidth: 150,
+            maxWidth: 200,
+            flex: 1,
+            type: 'singleSelect',
+            valueOptions: dishTypes,
+            valueFormatter({ value }: { value: string }) {
+                return capitalizeFirstLetter(value);
+            },
+            renderCell(params) {
+                const color = tagColor(params.value) ?? 'inherit';
+                return (
+                    <>
+                        {params && (
+                            <Chip
+                                variant="outlined"
+                                sx={{
+                                    color: `${color}`,
+                                    border: `2px solid ${color}`,
+                                }}
+                                label={params.formattedValue}
+                            />
+                        )}
+                    </>
+                );
+            },
+        },
+        {
+            field: 'transactionType',
+            headerName: 'Transaction Type',
+            minWidth: 150,
+            maxWidth: 200,
+            flex: 1,
+            type: 'singleSelect',
+            valueOptions: Object.values(TransactionType),
+            valueFormatter({ value }: { value: string }) {
+                return capitalizeFirstLetter(value);
+            },
+            renderCell(params) {
+                const color = tagColor(params.value) ?? 'inherit';
+                return (
+                    <>
+                        {params && (
+                            <Chip
+                                variant="outlined"
+                                sx={{
+                                    color: `${color}`,
+                                    border: `2px solid ${color}`,
+                                }}
+                                label={params.formattedValue}
+                            />
+                        )}
+                    </>
+                );
+            },
+        },
+        { field: 'userEmail', headerName: 'User', minWidth: 200, maxWidth: 250, flex: 1 },
+        {
+            field: 'timestamp',
+            headerName: 'Transaction Date',
+            minWidth: 200,
+            maxWidth: 300,
+            flex: 1,
+            type: 'date',
+            valueFormatter({ value }: { value: string }) {
+                return value ? new Date(value).toLocaleString() : null;
+            },
+        },
+    ];
 }
 
 export default function AdminTransactionsTable({ filteredRows, loadingTransactions, dishTypes }: Props) {
