+++ app/src/components/BottomTextInput.tsx
