+++ app/src/components/CameraScanner.tsx
