+++ app/src/components/QrScanner/hooks.tsx
