+++ app/src/components/QrScanner/index.tsx
