+++ app/src/components/QrScanner/styles.ts
