+++ app/src/components/QrScanner/types.ts
