+++ app/src/components/QrScanner/utils.ts
