+++ app/src/components/Scanner.tsx
